refactor: profileInfo분리
